Move ModelClassField into corehq.sql_db.fields
