+++ corehq/sql_db/fields.py
@@ -3,6 +3,43 @@ import sys
 from django.db.models import CharField
 
 
+class ModelClassField(CharField):
+    """Stores a slug in the DB but reads/writes Django model classes.
+
+    Lets callers query and assign with the class itself, e.g.
+    ``filter(model=XFormInstance)``.
+    """
+
+    def __init__(self, *args, **kwargs):
+        kwargs.setdefault('max_length', 128)
+        super().__init__(*args, **kwargs)
+
+    @property
+    def _slug_by_model(self):
+        from corehq.form_processor.models import CommCareCase, XFormInstance
+
+        return {CommCareCase: 'case', XFormInstance: 'xform'}
+
+    @property
+    def _model_by_slug(self):
+        return {v: k for k, v in self._slug_by_model.items()}
+
+    def from_db_value(self, value, expression, connection):
+        if value is None:
+            return value
+        return self._model_by_slug[value]
+
+    def to_python(self, value):
+        if value is None or value in self._slug_by_model:
+            return value
+        return self._model_by_slug[value]
+
+    def get_prep_value(self, value):
+        if value is None or isinstance(value, str):
+            return value
+        return self._slug_by_model[value]
+
+
 class CharIdField(CharField):
     """CharField that does not create varchar_pattern_ops index
